Refactor: Move UI files into 'screens' package
This commit refactors the project structure by moving all UI-related files (Screens, ViewModels, and UI components) into a dedicated `screens` subpackage within the `ui` package.

This change improves code organization and makes it easier to navigate the project.

Specifically, the following files and directories were moved:
- `app/src/main/java/com/kawunus/habityou/ui/root/` -> `app/src/main/java/com/kawunus/habityou/ui/screens/root/`
- `app/src/main/java/com/kawunus/habityou/ui/diary/` -> `app/src/main/java/com/kawunus/habityou/ui/screens/diary/`
- `app/src/main/java/com/kawunus/habityou/ui/newnote/` -> `app/src/main/java/com/kawunus/habityou/ui/screens/newnote/`
- `app/src/main/java/com/kawunus/habityou/ui/editnote/` -> `app/src/main/java/com/kawunus/habityou/ui/screens/editnote/`
- `app/src/main/java/com/kawunus/habityou/ui/usefulhabits/` -> `app/src/main/java/com/kawunus/habityou/ui/screens/usefulhabits/`
- `app/src/main/java/com/kawunus/habityou/MainActivity.kt` -> `app/src/main/java/com/kawunus/habityou/ui/MainActivity.kt`

The `ViewModelModule.kt` and `NavigationHost.kt` files were updated to reflect these changes.
`AndroidManifest.xml` was also updated to point to the new location of `MainActivity`.
A new `UsefulHabitsViewModel.kt` was created in the `app/src/main/java/com/kawunus/habityou/ui/screens/usefulhabits/viewmodel/` directory and added to the `ViewModelModule`.
