+++ app/src/main/java/com/kawunus/habityou/di/ViewModelModule.kt
@@ -1,9 +1,10 @@
 package com.kawunus.habityou.di
 
-import com.kawunus.habityou.ui.diary.viewmodel.DiaryViewModel
-import com.kawunus.habityou.ui.editnote.viewmodel.EditNoteViewModel
-import com.kawunus.habityou.ui.newnote.viewmodel.NewNoteViewModel
-import com.kawunus.habityou.ui.root.viewmodel.ToolbarViewModel
+import com.kawunus.habityou.ui.screens.diary.viewmodel.DiaryViewModel
+import com.kawunus.habityou.ui.screens.editnote.viewmodel.EditNoteViewModel
+import com.kawunus.habityou.ui.screens.newnote.viewmodel.NewNoteViewModel
+import com.kawunus.habityou.ui.screens.root.viewmodel.ToolbarViewModel
+import com.kawunus.habityou.ui.screens.usefulhabits.viewmodel.UsefulHabitsViewModel
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
 
@@ -16,4 +17,6 @@ val viewModelModule = module {
     viewModelOf(::NewNoteViewModel)
 
     viewModelOf(::EditNoteViewModel)
+
+    viewModelOf(::UsefulHabitsViewModel)
 }
