+++ app/src/main/java/com/kawunus/habityou/ui/MainActivity.kt
@@ -1,0 +1,22 @@
+package com.kawunus.habityou.ui
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import androidx.compose.material3.ExperimentalMaterial3Api
+import com.kawunus.habityou.ui.screens.root.ui.RootScreen
+import com.kawunus.habityou.ui.theme.HabituTheme
+
+class MainActivity : ComponentActivity() {
+    @OptIn(ExperimentalMaterial3Api::class)
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            HabituTheme {
+                RootScreen()
+            }
+        }
+    }
+}
