--- app/src/main/java/com/kawunus/habityou/ui/editnote/viewmodel/EditNoteScreenState.kt
@@ -1,10 +1,0 @@
-package com.kawunus.habityou.ui.editnote.viewmodel
-
-sealed class EditNoteScreenState {
-
-    object Edited : EditNoteScreenState()
-
-    object ReadyToEdit : EditNoteScreenState()
-
-    object Deleted : EditNoteScreenState()
-}
