--- app/src/main/java/com/kawunus/habityou/ui/newnote/ui/NewNotePreview.kt
@@ -1,13 +1,0 @@
-package com.kawunus.habityou.ui.newnote.ui
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.navigation.compose.rememberNavController
-
-@Composable
-@Preview
-fun NewNoteScreenPreview() {
-    NewNoteScreen(
-        navController = rememberNavController()
-    )
-}
