--- app/src/main/java/com/kawunus/habityou/ui/newnote/viewmodel/NewNoteScreenState.kt
@@ -1,8 +1,0 @@
-package com.kawunus.habityou.ui.newnote.viewmodel
-
-sealed class NewNoteScreenState {
-
-    object Done : NewNoteScreenState()
-
-    object ReadyToCreate : NewNoteScreenState()
-}
