+++ app/src/main/java/com/kawunus/habityou/ui/screens/diary/ui/DiaryEmpty.kt
@@ -1,0 +1,21 @@
+package com.kawunus.habityou.ui.screens.diary.ui
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import com.kawunus.habityou.R
+
+@Composable
+internal fun DiaryEmpty(modifier: Modifier = Modifier) {
+    Box(
+        modifier = modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ) {
+        Text(text = stringResource(R.string.diary_empty))
+    }
+}
+
