+++ app/src/main/java/com/kawunus/habityou/ui/screens/editnote/viewmodel/EditNoteScreenState.kt
@@ -1,0 +1,10 @@
+package com.kawunus.habityou.ui.screens.editnote.viewmodel
+
+sealed class EditNoteScreenState {
+
+    object Edited : EditNoteScreenState()
+
+    object ReadyToEdit : EditNoteScreenState()
+
+    object Deleted : EditNoteScreenState()
+}
