+++ app/src/main/java/com/kawunus/habityou/ui/screens/newnote/viewmodel/NewNoteScreenState.kt
@@ -1,0 +1,8 @@
+package com.kawunus.habityou.ui.screens.newnote.viewmodel
+
+sealed class NewNoteScreenState {
+
+    object Done : NewNoteScreenState()
+
+    object ReadyToCreate : NewNoteScreenState()
+}
