+++ app/src/main/java/com/kawunus/habityou/ui/screens/usefulhabits/viewmodel/UsefulHabitsViewModel.kt
@@ -1,0 +1,5 @@
+package com.kawunus.habityou.ui.screens.usefulhabits.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+class UsefulHabitsViewModel() : ViewModel()
